requirements.txt
# API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pgvector>=0.3.0

# Embeddings
sentence-transformers>=3.0.0

# Document parsing
pymupdf>=1.24.0
python-docx>=1.1.0

# Local LLM
ollama>=0.2.0

# Transcription
openai-whisper>=20231117

# Object storage
minio>=7.2.0

# Config
pydantic-settings>=2.2.0

# Utilities
httpx>=0.27.0
python-jose>=3.3.0
aiofiles>=23.2.0
